requirements.txt
asgiref==3.5.2
Brotli==1.0.9
cffi==1.15.1
cssselect2==0.7.0
Django==4.1.1
django-cors-headers==3.13.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
fonttools==4.37.3
html5lib==1.1
Jinja2==3.1.2
MarkupSafe==2.1.1
Pillow==9.2.0
pycparser==2.21
pydyf==0.3.0
PyJWT==2.5.0
pyphen==0.13.0
python-dotenv==0.21.0
pytz==2022.2.1
six==1.16.0
sqlparse==0.4.2
tinycss2==1.1.1
weasyprint==56.1
webencodings==0.5.1
xmltodict==0.13.0
zopfli==0.2.1
